feat: Add maxLength to dataModel
We were already collecting the maxLength for columns, we also already have the `maxLength` property on the type for data model scalar fields. We just needed to put them on the data model when converting from introspection result to data model for postgres (for sqlite there are no relevant char limits to consider).