feat: Support per-dialect shape logic (#19)
In the old repo, we would avoid calling the API if we already were able to determine a shape from just the type alone. This PR brings this back in the new repo, except in a per-dialect way.

This PR also centralises the logic for determining whether to generate a value for a field or not (e.g. if it is a generated id column or if it is an id sequence column), and reuses it in various places:
* in the plan when deciding whether to generate the value or not
* when determining which columns to pass to the predictions api
* when determining whether or not to generate a user model for the value as part of codegen