feat: Bring back passing fingerprint options to templates (#18)
For each field, depending on the field type we might generate options for it as part of the result of `snaplet fingerprint`. These options get passed through to each user model's generate function. In the old repo, our templates pass these options through to copycat. This PR is for having the templates support doing exactly this in this new repo.

<img width="845" alt="Screenshot 2024-03-06 at 11 20 43" src="https://github.com/snaplet/seed/assets/1731223/5f278140-beb4-44d2-9893-92e97d3680b2">


## Approach
Instead of using babel to augment the template as we did in the old repo, we have the template generate code that passes in the options to the copycat methods already. We do this by passing along extra context to each template, then having the `copycatTemplate` helper look for an `optionsInput` property in this extra context in order to determine whether or not an options argument has been passed in + what the relevant identifier's name is.