improvement: Get introspect command working with arbitrary dialects
Adds two new parts to the `Dialect` API:
* `getDataModel: (client: DrizzleDbClient) => Promise<DataModel>`
* `withDbClient: <Result>(props: { connectionString: string; fn: (client: DrizzleDbClient) => Promise<Result> | Result }) => Promise<Result>:

`getDataModel` is just an export of the already existing `getDataModel` we have for each dialect. `withDbClient` wraps up the boilerplate for getting our drizzle db client interface for each client, as well as taking care of tearing down the relevant db client.

We then use these new parts to make the `introspect` command work with arbitrary dialects.

With this in place, next step in new PR will be to get sqlite working with introspect command: https://linear.app/snaplet/issue/S-1927/fix-invalid-url-for-sqlite