fix: Do not use prediction API for non-string types (#48)
Co-authored-by: Carel de Waal <careldewaal@gmail.com>

The prediction API should only be used for string-based types. We have a function that we provide for each dialect (`determineShapeFromType`) that we use to decide on a shape by looking at the field type (i.e. it would not get sent to the API). This PR adds fixes to this function for each dialect so we only go to the API for string types, and if we could not find a shape for a non-string type, we use the `__DEFAULT` shape for it (to get the default code template).

Related discussion with context on the plan: https://discord.com/channels/788353076129038346/1219640183717105786