fix(cli): attempt connection before file watch
When we init snaplet for with some adapters (like prisma).

We might have to directly have a valid adapter connection there (because `new PrismaClient` will take care of loding the propre project configuration)

So we want to first check if the connection is already valid before starting our watcher over the config file for changes.