fix: Fix shape examples for the case of array types (#77)
When generating the code we use for model defaults, if the field type is an array type, we wrap it in `[]`s. Except we only do this for cases when we generate code from templates at the moment, and not also for the case where we use the shape examples.