fix: Save user id when talking to API for identifying users on posthog
There is some logic that was not ported over when switching to new seed repo (my bad): https://github.com/snaplet/snaplet/blob/bcf52cde1e6345fc25458f6ee2bab5d55304d51a/cli/src/lib/trpc.ts#L85

In this code, whenever we make API requests in the CLI, we get a user id response header, and save it to the system config. This means any subsequent posthog events we send would be associated with that user id.

Since we do not have this at the moment, we end up not being able to identify unique users if the user did not explicitly `login`, which is the case when users use `@snaplet/seed` on CI (with a `SNAPLET_ACCESS_TOKEN`).