fix: Save user id when talking to API for identifying users on posthog (#112)
There is some logic that was not ported over when switching to new seed repo (my bad): https://github.com/snaplet/snaplet/blob/bcf52cde1e6345fc25458f6ee2bab5d55304d51a/cli/src/lib/trpc.ts#L85

In this code, whenever we make API requests in the CLI, we get a user id response header, and save it to the system config. This means any subsequent posthog events we send would be associated with that user id.

Since we do not have this at the moment, we end up not being able to identify unique users if the user did not explicitly `login`, which is the case when users use `@snaplet/seed` on CI (with a `SNAPLET_ACCESS_TOKEN`).

## Note
There are some limitations to how effectively this solution will work in the new seed repo, because of the `generate` command. More context in these threads:
* https://discord.com/channels/788353076129038346/914821651017908234/1229426877320396880
* https://discord.com/channels/788353076129038346/914821651017908234/1229413193571635340

## Loom with a test run to confirm the fix works
https://www.loom.com/share/71b742074d2d47b5895a99d372b10f7d