fix: Fix telemetry production check (#133)
At the moment, we only fire telemetry in production (for e.g, to avoid spamming posthog with noise because of our own local dev and test runs, as well as CI runs).

The check we use for deciding whether it is production recently stopped working: we determine whether we are running in production by checking whether it is the source (`src`) files being run, or the built (`dist`) files. However, the way we were doing this was checking whether the parent dir's name was `src`, but a recent change made us have the dir structure for our built files be `./dist/src/*`.