chore: Remove production check for telemetry
We currently try determine if it is production to decide whether or not to send events to posthog. Thing is, there isn't a robust way of figuring out whether we are in production or not. We could explicitly say we are in dev (e.g. `NODE_ENV=development`), but it begs the question of why we need to do this. We already disable telemetry for our e2e tests (with